+++ stable_baselines3/__init__.py
@@ -12,3 +12,10 @@ from stable_baselines3.td3 import TD3
 version_file = os.path.join(os.path.dirname(__file__), "version.txt")
 with open(version_file, "r") as file_handler:
     __version__ = file_handler.read().strip()
+
+
+def HER(*args, **kwargs):
+    raise ImportError(
+        "Since Stable Baselines 2.1.0, `HER` is now a replay buffer class `HerReplayBuffer`.\n "
+        "Please check the documentation for more information: https://stable-baselines3.readthedocs.io/"
+    )