+++ docs/misc/changelog.rst
@@ -28,6 +28,7 @@ New Features:
   to handle gym3-style vectorized environments (@vwxyzjn)
 - Ignored the terminal observation if the it is not provided by the environment
   such as the gym3-style vectorized environments. (@vwxyzjn)
+- Added support for image observation when using ``HER``
 
 Bug Fixes:
 ^^^^^^^^^^